backend/pyproject.toml
[tool.poetry]
name = "copertine3"
version = "0.1.0"
description = ""
authors = ["Robert Alexander <gogonegro@gmail.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.59.3"
together = "^1.3.11"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
weaviate-client = "^4.10.2"
pydantic = "^2.10.4"
requests = "^2.32.3"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.25"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-mail = "^1.4.1"
python-multipart = "^0.0.9"
alembic = "^1.13.1"
click = "^8.1.7"
pyjwt = "^2.10.1"

[tool.poetry.scripts]
copertine-cli = "src.cli:cli"

[tool.poetry.group.test.dependencies]
playwright = "^1.41.1"
aiohttp = "^3.9.1"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"
mypy = "^1.14.1"
types-beautifulsoup4 = "^4.12.0.20241020"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
